requirements.txt
django==2.2
mysqlclient
pillow
redis
django-pure-pagination==0.3.0
django-simple-captcha